COMMERCE-1269 Rename package "data.store"
