+++ modules/etl/talend/src/main/java/com/liferay/talend/data/store/InputDataStore.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.talend.datastore;
+package com.liferay.talend.data.store;
 
 import java.net.URL;
